updated node version, functions and calls
