+++ database/firebase.json
@@ -3,7 +3,7 @@
       {
         "source": "functions",
         "codebase": "default",
-        "runtime": "nodejs18",
+        "runtime": "nodejs22",
         "ignore": [
           "node_modules",
           ".git",
