+++ database/functions/src/index.ts
@@ -5,40 +5,31 @@ import { scoreMatch } from "./scoreMatch.js";
 import { addParticipant } from "./addParticipant.js";
 import { fetchOrAddUser } from "./fetchOrAddUser.js";
 import { getCollege } from "./getCollege.js";
-import { getMatchesPaginated } from "./getMatchesPaginated.js";
 import { getUnscoredMatches } from "./getUnscoredMatches.js";
 import { getCollegeMatches } from "./getCollegeMatches.js";
 import { removeParticipant } from "./removeParticipant.js";
 import { getMatches } from "./getMatches.js";
-import { getSchedulePaginated } from "./getSchedulePaginated.js";
 import { updateUsername } from "./updateUsername.js";
-import { addBet } from "./addBet.js";
 import { deleteBet } from "./deleteBet.js";
-import { getPendingBets } from "./getPendingBets.js";
 import { getMyAvailablePoints } from "./getMyAvailablePoints.js";
 import { undoScoreMatch } from "./undoScoreMatch.js";
 import { publicApiSignup } from "./publicApis.js";
-import { addBetMod } from "./addBetMod.js";
-import { getBets } from "./getBets.js";
-import { getMatchesPaginatedTest } from "./getMatchesPaginatedTest.js";
 import { updateUserRole } from "./updateUserRole.js";
 import { getSeasons } from "./getSeasons.js";
 import { getSeasonPoints } from "./getSeasonPoints.js";
-import { getLeaderboardv2 } from "./getLeaderboardv2.js";
-import { getBetsv2 } from "./getBetsv2.js";
-import { getMatchesPaginatedv2 } from "./getMatchesPaginatedv2.js";
+import { getBets } from "./getBets.js";
+import { getMatchesPaginated } from "./getMatchesPaginated.js";
 import { createBracket } from "./createBracket.js";
 import { deleteBracket } from "./deleteBracket.js";
 import { scoreMatchTesting } from "./scoreMatchTesting.js";
-import { getSchedulePaginatedv2 } from "./getSchedulePaginatedv2.js";
 import { getUsersInCollege } from "./getUsersInCollege.js";
 import { assignRemoveCaptain } from "./assignRemoveCaptain.js";
 import { setMVP } from "./setMVP.js";
 import { addSchedule } from "./addSchedule.js";
-import { assignRemoveCaptainv2 } from "./assignRemoveCaptainv1.js";
+import { addBet } from "./addBet.js";
 
 export {
-  assignRemoveCaptainv2,
+  assignRemoveCaptain,
   getLeaderboard,
   getUserMatches,
   scoreMatch,
@@ -50,30 +41,21 @@ export {
   getCollegeMatches,
   removeParticipant,
   getMatches,
-  getSchedulePaginated,
   updateUsername,
   addBet,
   deleteBet,
-  getPendingBets,
   getMyAvailablePoints,
   getUserLeaderboard,
-  getLeaderboardv2,
   undoScoreMatch,
   publicApiSignup,
-  addBetMod,
   getBets,
-  getBetsv2,
-  getMatchesPaginatedTest,
   getSeasons,
   getSeasonPoints,
-  getMatchesPaginatedv2,
   updateUserRole,
   createBracket,
   deleteBracket,
   scoreMatchTesting,
-  getSchedulePaginatedv2,
   getUsersInCollege,
-  assignRemoveCaptain,
   setMVP,
   addSchedule,
 };
