+++ database/functions/src/publicApis.ts
@@ -3,6 +3,7 @@ import admin from "./firebaseAdmin.js";
 import cors from "cors";
 import jwt from "jsonwebtoken";
 import {randomBytes} from "crypto";
+import { SecretManagerServiceClient } from "@google-cloud/secret-manager";
 
 interface DecodedToken {
   name: string;
@@ -18,7 +19,6 @@ interface DecodedToken {
 const corsHandler = cors({ origin: true });
 
 const db = admin.firestore();
-const JWT_SECRET = "wp[rkfwprk0pki942j4e-9j1o[jfcm2lv;24p[2p0-1i0]]]";
 
 export const publicApiSignup = functions.https.onRequest((req, res) => {
   return corsHandler(req, res, async () => {
@@ -31,11 +31,15 @@ export const publicApiSignup = functions.https.onRequest((req, res) => {
         return res.status(401).json({error: "No token provided"});
       }
       const token = authHeader.split("Bearer ")[1];
-        
-      if (!JWT_SECRET) {
-        console.error("JWT_SECRET is not defined");
+      const client = new SecretManagerServiceClient();
+      const [version] = await client.accessSecretVersion({
+        name: "projects/yims-125a2/secrets/JWT_SECRET/versions/1",
+      });
+      if (!version.payload || !version.payload.data) {
+        console.error("JWT secret payload is missing");
         return res.status(500).json({error: "Internal Server Error"});
-                }
+      }
+      const JWT_SECRET = version.payload.data.toString();
 
       const decoded = jwt.verify(token, JWT_SECRET) as jwt.JwtPayload as DecodedToken;
