+++ frontend/src/app/api/functions/getBets/route.ts
@@ -11,7 +11,7 @@ export async function GET(req: Request) {
     const history = url.searchParams.get("history")
     const pending = url.searchParams.get("pending");
     const response= await fetch(
-        `https://getbetsv2-65477nrg6a-uc.a.run.app?seasonId=${seasonId}&history=${history}&pending=${pending}`,
+        `https://getbets-65477nrg6a-uc.a.run.app?seasonId=${seasonId}&history=${history}&pending=${pending}`,
         { headers: { "Content-Type": "application/json", Authorization: `Bearer ${token.value}` } }
     );
